The most significant changes involve the removal of custom media control files and the shift to using the standard `MediaElement` from the Maui toolkit. The `VideoPlayerPage.xaml` file was updated to replace the custom `MediaControl` with the `MediaElement`. The `VideoPlayerPage.xaml.cs` file was also updated to implement the `IRecipient<ShowItemMessage>` interface, suggesting that it can now receive messages of type `ShowItemMessage`. Additionally, new methods for handling media playback were added to this file.
List of changes:

1. The `MediaControl.xaml` and `MediaControl.xaml.cs` files were completely removed. These files contained the code for a custom media control that included various features such as play, pause, rewind, fast forward, mute, and full screen functionality.

2. The `NerdNewsNavigator2.csproj` file was updated to use version 8.0.7 of Maui instead of the previous 8.0.6-nightly.9851 version.

3. The `VideoPlayerPage.xaml` file was updated to replace the custom `MediaControl` with the `MediaElement` from the Maui toolkit.

4. The `VideoPlayerPage.xaml.cs` file was updated to implement the `IRecipient<ShowItemMessage>` interface. This file also includes new methods for handling media playback, such as `Seek`, `MediaStopped`, `IOSStart`, and `MediaControl_Unloaded`.
